+++ base/bootstrap.go
@@ -201,6 +201,10 @@ var _ mergo.Transformers = &mergoNilTransformer{}
 
 func (t *mergoNilTransformer) Transformer(typ reflect.Type) func(dst, src reflect.Value) error {
 	if typ.Kind() == reflect.Ptr {
+		if typ.Elem().Kind() == reflect.Struct {
+			// skip nilTransformer for structs, to allow recursion
+			return nil
+		}
 		return func(dst, src reflect.Value) error {
 			if dst.CanSet() && !src.IsNil() {
 				dst.Set(src)
